requirements.txt
# Core Dependencies
pandas>=2.0.0
pydantic>=2.0.0
sqlalchemy>=2.0.0

# LLM & Agent Framework
google-generativeai>=0.8.0
crewai>=0.86.0
crewai-tools>=0.14.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# UI
chainlit>=1.3.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0

# Utilities
python-dotenv>=1.0.0
rich>=13.0.0
